Turned in the project like this. More or less done.

+++ src/main/java/com/gemtastic/lillakammaren/model/Order.java
@@ -1,15 +1,17 @@
 package com.gemtastic.lillakammaren.model;
 
 /**
- *
+ * The order stores the three different object relevant to the orders made; the
+ * customer bean, the generated order number and the cart.
+ * 
  * @author Gemtastic
  */
 public class Order {
     private final Customer customer;
-    private final long orderno;
+    private final float orderno;
     private final Cart cart;
     
-    public Order(Customer customer, long orderno, Cart cart){
+    public Order(Customer customer, float orderno, Cart cart){
         this.customer = customer;
         this.orderno = orderno;
         this.cart = cart;
